requirements.txt
aiosqlite
anyio
asyncpg
fastapi
pydantic
pydantic_core
SQLAlchemy
uvicorn
loguru
python-dotenv
